fix(data): rename 'Leters' directory to 'Letters' and enhance documentation
